# app/docs/swagger.yaml
basePath: /v1
definitions:
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  model.OkLoginResponce:
    properties:
      email:
        type: string
      otp_enabled:
        type: boolean
      refresh_token:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  model.OkResponse:
    properties:
      message:
        type: string
    type: object
...
  model.ParamsRefreshTocken:
    properties:
      refresh_token:
        example: eyJhbGciOiJIU....FnjPC-zct_EDkIuUviRNI
        type: string
    type: object
...
  model.ResponceRefreshTocken:
    properties:
      access_token:
        example: eyJhbGciOiJIU....FnjPC-zct_EDkIuUviRNI
        type: string
      refresh_token:
        example: eyJhbGciOiJIU....FnjPC-zct_EDkIuUviRNI
        type: string
    type: object
  model.Track:
    properties:
      album:
        example: Album name
        type: string
      album_artist:
        example: Album artist name
        type: string
      artist:
        example: Artist name
        type: string
      bitrate:
        example: 320
        type: integer
      comment:
        example: Additional comments
        type: string
      composer:
        example: Composer name
        type: string
      disc:
        example: 1
        type: integer
      disc_total:
        example: 2
        type: integer
      duration:
        $ref: '#/definitions/time.Duration'
      genre:
        example: Genre name
        type: string
      lyrics:
        example: Lyrics of the track
        type: string
      sample_rate:
        example: 44100
        type: integer
      title:
        example: Title name
        type: string
      track:
        example: 3
        type: integer
      track_total:
        example: 10
        type: integer
      year:
        example: 2022
        type: integer
    type: object
  model.User:
    description: 'User account information with: user _id, name, email, password'
    properties:
      email:
        example: aaaa@aaaa.com
        type: string
      otp_enabled:
        type: boolean
      otp_verified:
        type: boolean
      otpAuthURL:
        type: string
      otpSecret:
        type: string
      password:
        example: "1111"
        type: string
    type: object
  model.UserResponse:
    properties:
      email:
        type: string
      name:
        type: string
      refresh_token:
        type: string
    type: object
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    - 1
    - 1000
    - 1000000
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
    - Nanosecond
    - Microsecond
    - Millisecond
...
host: s3streammedia.com
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a s3 media streamer server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: S3 Media Streamer Application API
  version: 0.0.1
paths:
...
  /tracks:
    get:
      consumes:
      - '*/*'
      description: responds with the list of all tracks as JSON.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Field to sort by (e.g., 'created_at')
        in: query
        name: sort_by
        type: string
      - description: Sort order ('asc' or 'desc')
        in: query
        name: sort_order
        type: string
      - description: Filter criteria ('I0001' or '=I0001')
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Track'
            type: array
        "400":
          description: Invalid page or page_size parameters
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Show the list of all tracks.
      tags:
      - track_handler-controller
  /tracks/{code}:
    get:
      consumes:
      - '*/*'
      description: |-
        locates the track_handler whose ID value matches the id parameter sent by the client,
        then returns that track_handler as a response.
      parameters:
      - description: Code track_handler
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Track'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Track whose ID value matches the id.
      tags:
      - track_handler-controller
  /users/delete:
    delete:
      consumes:
      - application/json
      description: Deletes the authenticated user_handler.
      produces:
      - application/json
      responses:
        "200":
          description: Success - User deleted
          schema:
            type: string
        "401":
          description: Unauthorized - User unauthenticated
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found - User not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Deletes a user_handler.
      tags:
      - user_handler-controller
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user_handler with provided email and password.
      parameters:
      - description: Login User
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/model.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/model.OkLoginResponce'
        "400":
          description: Bad Request - Incorrect Password
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found - User not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Authenticates a user_handler.
      tags:
      - user_handler-controller
  /users/logout:
    post:
      consumes:
      - application/json
      description: Clears the authentication cookie, logging out the user_handler.
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/model.OkResponse'
        "401":
          description: Unauthorized - User unauthenticated
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logs out a user_handler.
      tags:
      - user_handler-controller
  /users/me:
    get:
      consumes:
      - '*/*'
      description: Retrieves user_handler information based on JWT in the request's
        cookies
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/model.OkLoginResponce'
        "401":
          description: Unauthenticated
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user_handler information
      tags:
      - user_handler-controller
  /users/refresh:
    post:
      consumes:
      - application/json
      description: Validates the provided refresh token, generates a new access token,
        and returns it.
      parameters:
      - description: Refresh token
        in: body
        name: refresh_token
        required: true
        schema:
          $ref: '#/definitions/model.ParamsRefreshTocken'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully refreshed access token
          schema:
            $ref: '#/definitions/model.ResponceRefreshTocken'
        "400":
          description: Bad Request - invalid refresh token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid refresh token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: failed to generate tokens and cookies
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Refreshes the access token using a valid refresh token.
      tags:
      - user_handler-controller
  /users/register:
    post:
      consumes:
      - application/json
      description: Register a new user_handler with provided name, email, and password.
      parameters:
      - description: Register User
        in: body
        name: user_handler
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.UserResponse'
        "400":
          description: Bad Request - User with this email exists
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - User unauthenticated
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Registers a new user_handler.
      tags:
      - user_handler-controller
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"